Add macOS Windows portable and experimental shell artifacts
