+++ package.json
@@ -103,7 +103,9 @@
     "vnc:harness": "bun run runtime/scripts/vnc-harness-standalone.ts",
     "vnc:harness:report": "bun run runtime/scripts/playwright/vnc-harness-report.ts",
     "release:package-ux-reports": "bun run scripts/release/package-ux-reports.ts",
-    "release:build-linux-run": "bun run scripts/release/build-linux-run-artifact.ts"
+    "release:build-portable": "bun run scripts/release/build-portable-artifact.ts",
+    "release:build-linux-run": "bun run scripts/release/build-linux-run-artifact.ts",
+    "release:build-experimental-shell": "bun run scripts/release/build-experimental-shell-artifact.ts"
   },
   "dependencies": {
     "@fontsource-variable/inter": "5.2.8",
