+++ runtime/tsconfig.scripts.json
@@ -7,7 +7,9 @@
   },
   "include": [
     "../scripts/integration-test-suite.ts",
+    "../scripts/release/build-experimental-shell-artifact.ts",
     "../scripts/release/build-linux-run-artifact.ts",
+    "../scripts/release/build-portable-artifact.ts",
     "scripts/vnc-harness-standalone.ts",
     "scripts/playwright/oobe-local-container.ts",
     "scripts/playwright/vnc-harness-report.ts",
